Add a feature to synchronize the loaded shells and safety of shotguns
If there were more than 250 items on the ship and a client joined the lobby, the shotgun safety and number of shells loaded would not synchronize. To prevent this causing confusion, this feature adds a packet that will synchronize both to what the host sees when a shotgun is picked up.

Also, when changing the safety, only the owner of the shotgun will change the safety, and send a packet to the server to notify it and all other clients of the new setting.